+++ src/java/org/malbino/orion/facades/negocio/ProgramacionGruposFacade.java
@@ -41,7 +41,7 @@ public class ProgramacionGruposFacade {
     public boolean programarGrupos(GestionAcademica gestionAcademica, Carrera carrera, Nivel nivel, Turno turno, Integer capacidad) {
         List<Materia> materias = materiaFacade.listaMaterias(carrera.getId_carrera(), nivel);
         for (Materia materia : materias) {
-            Integer c1 = grupoFacade.cantidadGrupos(gestionAcademica.getId_gestionacademica(), materia.getId_materia()).intValue();
+            Integer c1 = grupoFacade.cantidadGrupos(gestionAcademica.getId_gestionacademica(), materia.getId_materia(), turno).intValue();
             String codigo = Constantes.ABECEDARIO[c1];
 
             Grupo grupo = new Grupo(codigo, capacidad, turno, true, gestionAcademica, materia);