+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.2] - 2021-08-27
+### Added
+### Changed
+### Removed
+### Fixed
+- Se corrigio el Historial Académico
+- Se corrigio el Centralizador Calificaciones
+- Se corrigio la programación de grupos
+
 ## [1.3.1] - 2021-08-26
 ### Added
 ### Changed